pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["brunobahri <brunobahri27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.0.1"
django = "^5.0.2"
django-rest-framework = "^0.1.0"
factory-boy = "^3.3.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.3.0"
psycopg2-binary = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
